modules/master/requirements.txt
azure-iothub-device-client~=1.4.3
requests
azure-storage-blob